# lib/resupply-api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title — orval uses it to name generated output files.
  title: Api
  version: 0.1.0
  description: |
    CPAP Resupply Automation admin API. Powers the admin dashboard
    (queue, episode detail, conversation viewer, fulfillment). Patient
    PHI flows through this API and lives in the resupply.* schema.
servers:
  - url: /resupply-api
    description: Base API path
...
paths:
...
  /sms/send-reminder:
    post:
      operationId: sendSmsReminder
      tags: [messaging]
      summary: Admin-initiated outbound SMS resupply reminder
      description: |
        Admin-initiated outbound SMS reminder. Looks up the patient
        + episode, opens a conversations row (channel=sms,
        status=open), upserts the inbound-lookup row in
        `phone_lookup`, and dispatches the message via Twilio
        Programmable Messaging.

        On success returns `{conversationId, messageSid}` — the
        patient phone number is NEVER returned to the dashboard.

        Returns 503 when SMS+Email is not configured (any of
        TWILIO_*, SENDGRID_*, RESUPPLY_PHONE_HMAC_KEY,
        RESUPPLY_LINK_HMAC_KEY missing). Feature-flagged on env
        presence — published behaviour.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendSmsReminderRequest"
      responses:
        "201":
          description: Reminder dispatched; Twilio is sending
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendSmsReminderResponse"
        "400":
          description: Body failed zod validation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessagingValidationError"
        "401":
          description: No valid session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthError"
        "403":
          description: Caller is not on the admin allowlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthError"
        "404":
          description: Patient or episode not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessagingError"
        "409":
          description: Patient is not in `active` status (paused, closed).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessagingError"
        "422":
          description: |
            Patient cannot be messaged — e.g. no phone on file, or
            episode does not belong to this patient, or phone is not
            a valid E.164 number.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessagingError"
        "502":
          description: Twilio rejected the send.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessagingError"
        "503":
          description: |
            SMS+Email subsystem is not configured on this server.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessagingError"
...
  /rules:
    get:
      operationId: listRules
      tags: [rules]
      summary: List frequency rules (active + inactive)
      description: |
        Returns every rule, ordered by (priority asc, createdAt asc) —
        the same order the eligibility engine uses. Inactive rules are
        included so admins can re-enable them without re-entering
        the criteria.
      responses:
        "200":
          description: Rules list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FrequencyRuleList"
        "401":
          description: No session.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthError"
        "403":
          description: Caller is not on the admin allowlist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthError"
    post:
      operationId: createRule
      tags: [rules]
      summary: Create a frequency rule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FrequencyRuleCreate"
      responses:
        "201":
          description: Rule created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FrequencyRule"
        "400":
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsoleValidationError"
        "401":
          description: No session.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthError"
        "403":
          description: Caller is not on the admin allowlist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthError"
  /rules/{id}:
    patch:
      operationId: updateRule
      tags: [rules]
      summary: Update a frequency rule
      description: |
        Partial update. Any field omitted from the body is left alone;
        nullable fields can be cleared by sending an explicit `null`.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FrequencyRuleUpdate"
      responses:
        "200":
          description: Update applied.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FrequencyRuleUpdateResponse"
        "400":
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsoleValidationError"
        "404":
          description: No rule with that id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
        "401":
          description: No session.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthError"
        "403":
          description: Caller is not on the admin allowlist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthError"
    delete:
      operationId: deleteRule
      tags: [rules]
      summary: Delete a frequency rule
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Deleted.
        "404":
          description: No rule with that id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
        "401":
          description: No session.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthError"
        "403":
          description: Caller is not on the admin allowlist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthError"
...
components:
  schemas:
...
    AuthError:
      type: object
      description: |
        Auth failure body. Carries a single human-readable `error`
        string safe for the dashboard to render verbatim. Never
        contains the underlying error message, the user id, or any
        connection details.
      properties:
        error:
          type: string
      required:
        - error
...
    SendSmsReminderRequest:
      type: object
      description: |
        Admin instruction to send an outbound SMS reminder.
      properties:
        patientId:
          type: string
          format: uuid
          description: Patient row to message.
        episodeId:
          type: string
          format: uuid
          description: |
            Episode the reminder is associated with. Optional — if
            omitted, the most recent episode for this patient is
            used. When provided, must belong to patientId.
        body:
          type: string
          minLength: 1
          maxLength: 1600
          description: |
            Optional override for the SMS body. When omitted, the
            server renders the default reminder template using the
            patient's first name.
      required:
        - patientId
      additionalProperties: false
    SendSmsReminderResponse:
      type: object
      description: |
        Returned on a successful SMS reminder dispatch. Contains only
        opaque identifiers — never the patient phone number.
      properties:
        conversationId:
          type: string
          format: uuid
          description: New conversations row id.
        messageSid:
          type: string
          description: Twilio MessageSid for cross-referencing with Twilio logs.
      required:
        - conversationId
        - messageSid
...
    MessagingError:
      type: object
      description: |
        Error body returned by SMS+Email endpoints. Carries a stable
        machine-readable `error` code and an optional human-readable
        `message`. NEVER carries vendor error text or PHI.
      properties:
        error:
          type: string
          description: Stable error code.
          enum:
            - messaging_not_configured
            - patient_not_found
            - patient_not_active
            - patient_missing_phone
            - patient_missing_email
            - patient_phone_unnormalizable
            - episode_not_found
            - episode_patient_mismatch
            - no_episode_for_patient
            - conversation_create_failed
            - twilio_config_error
            - twilio_api_error
            - sendgrid_config_error
            - sendgrid_api_error
        message:
          type: string
        twilioStatus:
          type: integer
        twilioCode:
          type: string
        sendgridStatus:
          type: integer
      required:
        - error
    MessagingValidationError:
      type: object
      description: |
        Returned when an SMS+Email request body fails zod validation.
      properties:
        error:
          type: string
          enum: [invalid_body]
        issues:
          type: array
          items:
            type: object
            properties:
              path:
                type: string
              message:
                type: string
            required: [path, message]
      required:
        - error
        - issues
...
    ConsoleValidationError:
      type: object
      description: Returned when a list endpoint's query string fails zod validation.
      properties:
        error:
          type: string
          enum: [invalid_query]
        issues:
          type: array
          items:
            type: object
            properties:
              path:
                type: string
              message:
                type: string
            required: [path, message]
      required:
        - error
        - issues
    NotFoundError:
      type: object
      description: Returned when a {id} path parameter does not match any row.
      properties:
        error:
          type: string
          enum: [not_found]
      required:
        - error
...
    FrequencyRule:
      type: object
      description: |
        A frequency rule consulted by the eligibility engine when a
        patient has no per-patient override. NULL `match_*` predicates
        mean "this rule does not constrain on that axis".
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        priority:
          type: integer
          minimum: 0
          description: Lower priority is evaluated first.
        matchItemSkuPrefix:
          type: string
          nullable: true
        matchInsurancePayer:
          type: string
          nullable: true
        minTenureDays:
          type: integer
          nullable: true
          minimum: 0
        maxTenureDays:
          type: integer
          nullable: true
          minimum: 0
        cadenceDays:
          type: integer
          minimum: 1
          maximum: 365
        defaultChannel:
          type: string
          nullable: true
          enum: [sms, email, voice]
        active:
          type: boolean
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - priority
        - matchItemSkuPrefix
        - matchInsurancePayer
        - minTenureDays
        - maxTenureDays
        - cadenceDays
        - defaultChannel
        - active
        - notes
        - createdAt
        - updatedAt
    FrequencyRuleList:
      type: object
      properties:
        rules:
          type: array
          items:
            $ref: "#/components/schemas/FrequencyRule"
      required:
        - rules
    FrequencyRuleCreate:
      type: object
      description: |
        Body for POST /rules. `name` and `cadenceDays` are required;
        every other field is optional or nullable per the same
        semantics as PATCH.
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        priority:
          type: integer
          minimum: 0
          maximum: 100000
        matchItemSkuPrefix:
          type: string
          nullable: true
          maxLength: 120
        matchInsurancePayer:
          type: string
          nullable: true
          maxLength: 120
        minTenureDays:
          type: integer
          nullable: true
          minimum: 0
          maximum: 36500
        maxTenureDays:
          type: integer
          nullable: true
          minimum: 0
          maximum: 36500
        cadenceDays:
          type: integer
          minimum: 1
          maximum: 365
        defaultChannel:
          type: string
          nullable: true
          enum: [sms, email, voice]
        active:
          type: boolean
        notes:
          type: string
          nullable: true
          maxLength: 2000
      required:
        - name
        - cadenceDays
      additionalProperties: false
    FrequencyRuleUpdate:
      type: object
      description: |
        Body for PATCH /rules/{id}. Every field is optional; sending
        `null` explicitly clears a nullable field.
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        priority:
          type: integer
          minimum: 0
          maximum: 100000
        matchItemSkuPrefix:
          type: string
          nullable: true
          maxLength: 120
        matchInsurancePayer:
          type: string
          nullable: true
          maxLength: 120
        minTenureDays:
          type: integer
          nullable: true
          minimum: 0
          maximum: 36500
        maxTenureDays:
          type: integer
          nullable: true
          minimum: 0
          maximum: 36500
        cadenceDays:
          type: integer
          minimum: 1
          maximum: 365
        defaultChannel:
          type: string
          nullable: true
          enum: [sms, email, voice]
        active:
          type: boolean
        notes:
          type: string
          nullable: true
          maxLength: 2000
      additionalProperties: false
    FrequencyRuleUpdateResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        changed:
          type: array
          items:
            type: string
      required:
        - id
        - changed
...